+++ duckscript_sdk/src/sdk/std/fs/is_file/mod.rs
@@ -1,0 +1,45 @@
+use crate::utils::pckg;
+use duckscript::types::command::{Command, CommandResult};
+use fsio::path::as_path::AsPath;
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "IsFile")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["is_file".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn run(&self, arguments: Vec<String>) -> CommandResult {
+        if arguments.is_empty() {
+            CommandResult::Error("Path not provided.".to_string())
+        } else {
+            let path = &arguments[0].as_path();
+            CommandResult::Continue(Some(path.is_file().to_string()))
+        }
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}