+++ duckscript_sdk/src/sdk/std/fs/mod.rs
@@ -3,6 +3,8 @@ mod basename;
 mod canonical;
 mod cp;
 mod dirname;
+mod is_directory;
+mod is_file;
 mod list;
 mod mkdir;
 mod mv;
@@ -30,6 +32,8 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(canonical::create(&package))?;
     commands.set(cp::create(&package))?;
     commands.set(dirname::create(&package))?;
+    commands.set(is_directory::create(&package))?;
+    commands.set(is_file::create(&package))?;
     commands.set(list::create(&package))?;
     commands.set(mkdir::create(&package))?;
     commands.set(mv::create(&package))?;